Add the brew installer
The only problem is now, we'll have to get sudo access to create /srv
for the server files.  I'd like to put it there for various reasons
including the ability to move the admin app around and have it still
work.  I though about putting it in ~/Library but the username can
sometimes have spaces and some packages might not like it.  If i can
think of something better, we'll switch.